+++ src/resolver/apidom/reference/dereference/strategies/openapi-3-1-swagger-client/visitors/schema-nested/properties.js
@@ -1,6 +1,6 @@
 import { isObjectElement, toValue } from '@swagger-api/apidom-core';
 
-import toPath from '../utils/to-path.js';
+import toPath from '../../utils/to-path.js';
 
 class ModelPropertyMacroVisitor {
   modelPropertyMacro;
